requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.7.1
python-dotenv==1.0.1
python-telegram-bot==21.0.1
groq==0.8.0
python-multipart==0.0.9
httpx==0.27.0
aiofiles==23.2.1
